Created empty arena
Created an empty arena that is a white background with invisible walls that prevent the player from leaving the playable area. The walls are set up using staticbody2d node.

Also added the arena to the DELETEMETEST scene, so that you can see that the player cannot escape the bounds.