Damage Element and Static Damage Element complete
Worked on bounce behaviour as a group, changed from adding velocity to setting velocity.

Added bounce_force as export variable, as this is the set velocity.

If vector is zero (edge case), it gets difference in locations to find direction.

Also changed layers for walls to make sure the player collides with them.