Started setting up GameManager scene
GameManager is an autoload scene that allows data to persist between scenes in Godot. It currently holds values for the selected characters, their health values, and the current round number.

The lists of all characters and unlocked characters will be condensed into a dictionary.

Started experimenting with generating our own documentation in Godot's Help menu.